+++ src/di-container.ts
@@ -1,6 +1,8 @@
 import { Container } from 'inversify';
 import { IMarkdown } from './markdown/markdown.interface';
 import { MarkdownService } from './markdown/markdown.service';
+import { ITocService } from './toc/toc.interface';
+import { TocService } from './toc/toc.service';
 import { TYPES } from './types';
 
 /**
@@ -11,9 +13,7 @@ export class DiContainer {
 
   constructor() {
     this.diContainer = new Container();
-    this.diContainer
-      .bind<IMarkdown>(TYPES.MarkdownService)
-      .to(MarkdownService)
-      .inSingletonScope();
+    this.diContainer.bind<IMarkdown>(TYPES.MarkdownService).to(MarkdownService).inSingletonScope();
+    this.diContainer.bind<ITocService>(TYPES.TocService).to(TocService).inSingletonScope();
   }
 }