+++ src/toc/toc.ts
@@ -2,10 +2,10 @@
 
 import 'reflect-metadata';
 import { inject, injectable } from 'inversify';
-import { IHeading } from '../models/toc.interface';
-import { toKebabCase } from '../utils/utils';
+import { IHeading, IValidation } from '../models/toc.interface';
+import { Color, log } from '../utils/utils';
 import { IMarkdown } from '../markdown/markdown.interface';
-import { IToc } from './toc.interface';
+import { IToc, ITocService } from './toc.interface';
 import { TYPES } from '../types';
 
 /**
@@ -25,24 +25,27 @@ export class Toc implements IToc {
   private _mdPath = '';
   private _mdString = '';
 
-  private readonly _tocPlaceholder = new RegExp(
-    '<!--\\s?toc\\s?-->\n?(?<toc>(?:.*\n)+)<!--\\s?tocstop\\s?-->\n'
-  );
+  private readonly _tocPlaceholder = new RegExp('<!--\\s?toc\\s?-->\n?(?<toc>(?:.*\n)+)<!--\\s?tocstop\\s?-->\n');
   private readonly _tocStart = '<!-- toc -->\n';
   private readonly _tocStop = '<!-- tocstop -->\n';
   private readonly _indentation = '  ';
   private readonly _hyphen = '- ';
 
+  /** file path of markdown file */
   public get filePath(): string {
     return this._mdPath;
   }
 
+  /** set file path of markdown file */
   public set filePath(path: string) {
     this._mdPath = path;
-    this._mdString = this._mdService.parseMarkdown(path);
+    this._mdString = this.mdService.parseMarkdown(path);
   }
 
-  constructor(@inject(TYPES.MarkdownService) private _mdService: IMarkdown) {}
+  constructor(
+    @inject(TYPES.MarkdownService) public mdService: IMarkdown,
+    @inject(TYPES.TocService) public tocService: ITocService
+  ) {}
 
   /**
    * create table of content
@@ -50,17 +53,15 @@ export class Toc implements IToc {
    */
   public createToc(): string {
     let toc = '';
-    const cleandedContent = this._mdService.removeCodeBlocks(this._mdString);
+    const cleandedContent = this.mdService.removeCodeBlocks(this._mdString);
 
-    this._mdService
-      .parseHeadings(cleandedContent)
-      .forEach((heading: IHeading) => {
-        toc +=
-          this._indentation.repeat(heading.level - 1) +
-          this._hyphen +
-          this.createTocEntry(heading.heading, heading.counter) +
-          '\n';
-      });
+    this.mdService.parseHeadings(cleandedContent).forEach((heading: IHeading) => {
+      toc +=
+        this._indentation.repeat(heading.level - 1) +
+        this._hyphen +
+        this.createTocEntry(heading.heading, heading.counter) +
+        '\n';
+    });
 
     return toc;
   }
@@ -71,11 +72,8 @@ export class Toc implements IToc {
   public insertToc(): void {
     const tocMatch = this._tocPlaceholder.exec(this._mdString);
     if (tocMatch && tocMatch[0]) {
-      const mdWithToc = this._mdString.replace(
-        tocMatch[0],
-        this._tocStart + this.createToc() + this._tocStop
-      );
-      this._mdService.updateMarkdown(this.filePath, mdWithToc);
+      const mdWithToc = this._mdString.replace(tocMatch[0], this._tocStart + this.createToc() + this._tocStop);
+      this.mdService.updateMarkdown(this.filePath, mdWithToc);
       return;
     }
     throw new Error(
@@ -90,7 +88,35 @@ export class Toc implements IToc {
    * set max depth for parsing headings
    */
   public setMaxDepth(maxDepth: number): void {
-    this._mdService.setMaxDepth(maxDepth);
+    this.mdService.setMaxDepth(maxDepth);
+  }
+
+  /**
+   * evaluates if given toc is valid
+   * the given toc will be compared with the generated toc
+   * @returns is table of content valid
+   */
+  public isTocValid(): boolean {
+    const cleandedContent = this.mdService.removeCodeBlocks(this._mdString);
+    const headings: IHeading[] = this.mdService.parseHeadings(cleandedContent);
+    const headingValidation: IValidation = this.tocService.validateToc(cleandedContent, headings);
+
+    if (headingValidation.existingHeadingsValidation.length === 0) {
+      log(`[ERR] Couldn't find a table of content (${this._mdPath})`, Color.RED);
+      return false;
+    }
+
+    if (
+      headingValidation.existingHeadingsValidation.filter(
+        (item) => !item.validCaption || !item.validLevel || !item.validLink
+      ).length ||
+      headingValidation.missingHeadingToc.length
+    ) {
+      log(`[ERR] table of content is outdated (${this._mdPath})`, Color.RED);
+      return false;
+    }
+
+    return true;
   }
 
   /**
@@ -100,9 +126,6 @@ export class Toc implements IToc {
    * @returns toc entry
    */
   private createTocEntry(caption: string, counter: number): string {
-    if (counter) {
-      return `[${caption}](#${toKebabCase(caption)}-${counter})`;
-    }
-    return `[${caption}](#${toKebabCase(caption)})`;
+    return `[${caption}]${this.tocService.createLink(caption, counter)}`;
   }
 }