+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 # CHANGELOG
 
+## 0.2.0 (soon)
+
+- add new logo
+- format code with prettier
+- refactor app and exclude logic in external service
+- add maxDepth option to define max parsed heading level
+- add `check` command to check if given toc is valid and exists
+
 ## 0.1.0 (13.02.2021)
+
 - init project